compute grad

+++ makemore/src/data.rs
@@ -2,18 +2,18 @@ use anyhow;
 use std::collections::HashMap;
 use nalgebra::*; 
 
-pub fn build_dataset(words: Vec<&str>, block_size: usize, stoi: &HashMap<char, usize>) -> anyhow::Result<(DMatrix<f32>, DVector<f32>)> {
-    let mut x: Vec<Vec<f32>> = Vec::new();
-    let mut y: Vec<f32> = Vec::new();
+pub fn build_dataset(words: Vec<&str>, block_size: usize, stoi: &HashMap<char, usize>) -> anyhow::Result<(DMatrix<f64>, DVector<f64>)> {
+    let mut x: Vec<Vec<f64>> = Vec::new();
+    let mut y: Vec<f64> = Vec::new();
 
     words.iter().for_each(|w| {
         let mut context = vec![0.0; block_size];
         w.chars().chain(".".chars()).for_each(|ch| {
             let ix = stoi[&ch];
             x.push(context.clone());
-            y.push(ix as f32);
+            y.push(ix as f64);
             context = context[1..].to_vec();
-            context.push(ix as f32);
+            context.push(ix as f64);
         });
     });
